// functions/src/schemas/medication.schema.ts
export interface MedicationSchema {
  medName: string | Object | undefined;
  aroundMeal: string | Object | undefined;
  // durationAmount: number;
  // durationUnit: string;
  perDose: string | Object | undefined;
  perDay: string | Object | undefined;
  doseUnit: string | Object | undefined;
}

// export interface Duration {
//   unit: string;
//   amount: number;
// }